adding systematic sampling selection
adding an algorithm for systematic sampling selection
that follows the statistical approach of systematic sampling.
This commit introduces a simple factory that creates
the correct implementation. Furthermore, the user can
select the desired representative samples algorithm based
on a predefined property. This should be potentially
changed to be dynamic user input in the cmd as well as
the web app.