fixing collection selection [REF #21]
changin the click() event handler with change()
event handler. In order to allow a change the
getCollections method of the Application controller
always appends an empty collection at the beginning.

Also fixing some minor unnecessary checks and logs
in the javascript function.

Fixing a bug in the Application controller REST API
that returns xml instead of json