enhancing the adaptor and gatherer interface
improving the adaptor interface. Now the implementor does
not have to bother with the controller reference and storing the
data to the persistence layer. The only thing that one has to do
is to parse the actual data. All the logic is implemented
in the abstract class.

Adding a prototype of a new LocalFileGatherer, that will be
thread based.

Adding some javadocs