Cleaning up repository
Removing many deprecated methods that are not needed anymore.
Removing unnecessary constants and methods.

This commit marks a stage where the backend abstracted (fixes #46),
the filtering is enhanced (fixes #11) with a new convention.
Refactoring of the profile genration (fixes #48) as well as
the sampling algorithms (fixes #49) is also done.
The Controller allows data consolidation (fixes #47) based
on the identifier of an element.
The whole CLI is refactored and is much cleaner as it
uses a different library, but also does not violate the
architecture of C3PO and uses only the Configurator and Controller
directly (fixes #50).
All current tests are running successfully (fixes #34)